[dev] Add Modes to Weapon
This is how the weapon works:
You have 3 shooting modes, you can change with the mouse wheel, or through the numbers 1,2 and 3.

Pull Mode: allows you to attract an object when you click the primary button, if you click again it releases it. If you change mode while an object is held, it remains held for 3 seconds (to be able to launch it with the push mode, or freeze it with the freeze mode), if the 3 seconds pass, this object is released.

Push Mode: if there is no object attached, push in the area. If there is an object held, he throws it.

Freeze Mode: freezes an object in the air for a while.